7/24/25  Daily Update 1
Reworked the screen sector system to use pre-defined angles to differentiate the sectors, rather than coordinates. This is a far better approach than what I had previously, where I was defining the sectors as polygons, and was going to have to calculate if the mouse pointer was in the polygon in order to determine which screen sector it was in. With my new approach I only have to see if the angle between the player's center point and the mouse pointer is within the bounds of the set of angles that define the screen sectors.  credit goes to Professor Taylor for suggesting this approach.

As a part of the prior change, I also reworked the debug function displaySectors to display the sectors using the updated angle system.

I also changed the origin point of from which the angles are referenced from the player's head to the player's center, so as to better match the positions of the shield spell.
I also added a simple "cancel" combo to un-charge whatever spell is currently charged, but not cast.  Along with the combo I also added another bit of GUI text like 'OVERCHARGED' and 'COMBO FAILED' to indicate that the currently charged spell has been cancelled. In adding this new combo splash text, I reworked the method in the drawCombo function by which the code determines which image to display depending on the current value of the app.displayCombo variable. It used to rely on a series of conditionals, but I reworked it to use a dictionary instead, where each image path is matched to it's respective expected value of app.displayCombo.

Along with the screen sector rework I added a debug feature to display what screen sector your mouse is in on every click, to ensure that my detection method was working correctly.

I also implemented the very beginnings of the shield spell code, getting the appropriate image to display on screen when the spell is cast. Tomorrow I hope to actually implement the hitbox initiation, as well as work through proper angled rectangle collisions.

Some minor additional changes was re-organizing some of my code segmentations, and adjusting their spacings. I also drew the pause text, and worked it into my GUI