7/25/25 - Daily Update 2
added a shield removal feature, where if the shield is cast you can immediately uncast it by right clicking.

I added the angled rectangle collision system, which took considerable effort and assistance from Professor Taylor and also TA Nathan Xie. My approach is referenced from the website https://www.gorillasun.de/blog/an-algorithm-for-polygon-intersections/, but I put a lot of work into understanding exactly how the algorithm they used works, and made sure to detail my understanding of it's functionality in the code. (see checkConvexPolygonIntersection) I also had to re-do the ordering of the verticies for my recatngle and angledRectangle hitboxes to make them work with the mathematical methods used in this approach.

Also fixed a bug where if an aggressive spell was charged while the shield spell was active it would rerun the call to get the current mouse screen sector and re-orient the shield without re-casting the spell.

Minor changes - I noticed that I swapped left and right a lot for my screen sector system, I fixed that. I redid the order of the strings in the screen sector section, as well as the screen angle setup, and renamed the associated shield image files. I also added a small animation feature, and used it to display an ending animation for level one slime ball.