7/26/25 - Daily Update 3
Started working on the final boss. I drew two images for it and it's title, and worked it into my GUI and basic game loop. It now appears on game start rather than the Hideous Beast. I added an awakening feature, where it appears dormant at first, but awakens when the player deals damage to it. this also causes it's boss bar to appear on screen. I also made it rise from the ground, and follow the player.

I also reworked the hitbox collision effect system to allow collisions with more than just The Hideous Creature, as before that was the only enemy that the hitbox collision effect system could handle.

Tomorrow I intend to make the boss be able to actually shoot projectiles at the player, as right now the only way it can damage the player is if the player runs into it, which isn't a very enjoyable boss fight :/

I also updated the angled rectangle and circle collision to also check the distances between any of the verticies of the angled rectangle and the center of the circle

I also changed the combo for ghost Sword, and updated the spell scroll sheet to include both the ghost sword and the cancel combo