7/28/25 Daily Update 4
made the Insidious Creature boss able to fire projectiles, but also not just that, but MULTIPLE PROJECTILES simultaneously, which is something the wizard cannot do. I reworked the projectile initiation system to still work with multiple kinds of the same projectile, but with different names.

I also added the beginnings of my phase system to the Insidious Creature boss, and upped it's max health from 500 to 1000.

also added magic bounding boss around the insidious creature, so the player can't just run away and charge up while doing so.  The bounding box damages the player when they are outside of it.