7/28/25 Daily Update 5
Projectiles and boss can now collide with the shield, which deletes instantly on contact with the boss, and is progressively damaged on contact with projectiles until broken. I also redrew the images for the shield so that the apparent size is consistent at all angles. Also I wanted them to be prettier and not scribbly ms paint drawings.

I also added an out-of-bounds feature, where the screen fades to black and the player dies if their overall x coordinate surpasses a set threshold.

I also changed the way by which the bounding box around the boss detects whether or not the player has moved outside of it.

I also added an overlay below the initial spawning position of the insidious creature

I drew the image for phase 2 of the boss, and added a system for health and mana pickups. It isn't completely done yet, but it spawns them and initiates their hitboxes.