7/30/2025 Daily Update 7
I added a boss death sequence and a win condition. Killing the boss makes him explode, and plays a cutscene where you receive your hard-earned wizarding license, and after a while, it prompts you to restart the game by pressing 'r'. I also fixed the bug where the projectiles in the boss's final attack spawn in an incorrect location. It necessitated me pulling apart my entire hitbox and linear projectile system, but eventually worked.

I now consider the game to be in a 'finished' state. I do not like that the only really usable projectile is the ghost sword, but I don't really have time to add more spells. I will try to do so, but I suspect that the video will be difficult to make, and will end up eating a lot of my time.

I also did a ton of organization and explaining, to try to make my code as clear as possible.