7/31/25 Daily Update 7.2
Fixed a bug where I forgot to transition the subphase for the insidious creature from phaseChange_3 to something else, which was making the boss permanently stay on the phase two image during it's third phase. This bug is visible in the video, but it is far too much effort to redo as of right now, and barely noticeable. I just worked too hard on that final sprite for it not to be used, so fixing the code it is. perhaps it is bonus content :3