7/31/2025 Daily Update 8
Added one more spell, the cool lightning one I've been wanting to add since the very beginning.  I also cleaned up the projectile update system for the insidious creature's projectiles, and created an alternative arrangement of the code that adheres to the 80 character limit, for the sake of style, and also as a chance to review my code and it's behavior prior to the final submission.